requirements.txt
# API
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.17
pydantic==2.10.3
pydantic-settings==2.6.1

# ML (versions aligned for modern Python; adjust if your platform pins differ)
torch>=2.6.0
torchvision>=0.21.0
transformers==4.46.3
safetensors==0.4.5
accelerate==0.34.2
datasets==3.1.0
numpy==1.26.4
Pillow==10.4.0
scikit-learn==1.5.2

# Optional narrative quality (install if used)
openai==1.57.4
